refactor(frame): reset the structure of ethereum module
