+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "darwinia-bridge-primitives"
-version = "0.0.7"
+version = "0.0.8"
 authors = ["clearloop <udtrokia@gmail.com>"]
 description = "Darwinia bridge primitives"
 repository = "https://github.com/darwinia-network/bridge-primitives"
