+++ src/frame/ethereum/game.rs
@@ -7,11 +7,11 @@ use substrate_subxt_proc_macro::{module, Store};
 
 /// Ethereum Relay Pallet
 #[module]
-pub trait RelayerGame: System {}
+pub trait EthereumRelayerGame: System {}
 
 /// PendingHeaders Storage
 #[derive(Clone, Debug, Eq, PartialEq, Store, Encode)]
-pub struct PendingHeadersStore<T: RelayerGame> {
+pub struct PendingHeadersStore<T: EthereumRelayerGame> {
     #[store(returns = Vec<PendingHeader>)]
     /// Runtime marker
     pub _runtime: PhantomData<T>,
