+++ src/frame/ethereum/mod.rs
@@ -1,4 +1,5 @@
 //! Darwinia Frame
 #![cfg(feature = "runtime")]
 
-pub mod ethereum;
+pub mod game;
+pub mod relay;
